refactor(territory): merge import logic into territory and drop content-import module
- Relocate `DivisionImportService`, `DivisionFeatureMapper`, and related documentation to the `territory` module
- Move `ImportController` to the `pt.estga.territory.controllers` package
- Delete the now-obsolete `content-import` module and its Spring Boot auto-configuration
- Remove `content-import` from the root POM's modules and dependency management
- Remove `content-import` dependency from the `boot` module
