--- content-import/src/main/java/pt/estga/contentimport/config/ContentImportAutoConfiguration.java
@@ -1,9 +1,0 @@
-package pt.estga.contentimport.config;
-
-import org.springframework.boot.autoconfigure.AutoConfiguration;
-import org.springframework.context.annotation.ComponentScan;
-
-@AutoConfiguration
-@ComponentScan("pt.estga.contentimport")
-public class ContentImportAutoConfiguration {
-}
